feat: split modules into directories and open repos popup with Ctrl+d
